refactor: move compressor charts into common
