+++ src/libecalc/dto/models/compressor/chart.py
@@ -3,11 +3,10 @@ from typing import Union
 from pydantic import Field
 from typing_extensions import Annotated
 
+from libecalc.common.serializable_chart import SingleSpeedChartDTO, VariableSpeedChartDTO
 from libecalc.dto.models.chart import (
     GenericChartFromDesignPoint,
     GenericChartFromInput,
-    SingleSpeedChartDTO,
-    VariableSpeedChartDTO,
 )
 
 CompressorChart = Annotated[
