+++ src/libecalc/dto/models/compressor/stage.py
@@ -3,8 +3,9 @@ from typing import List, Optional
 from pydantic import Field
 from typing_extensions import Annotated
 
+from libecalc.common.serializable_chart import VariableSpeedChartDTO
 from libecalc.dto.base import EcalcBaseModel
-from libecalc.dto.models.compressor.chart import CompressorChart, VariableSpeedChartDTO
+from libecalc.dto.models.compressor.chart import CompressorChart
 from libecalc.dto.types import FixedSpeedPressureControl
