+++ src/libecalc/fixtures/compressor_process_simulations/compressor_process_simulations.py
@@ -4,6 +4,7 @@ from typing import NamedTuple
 import pytest
 
 from libecalc import dto
+from libecalc.common.serializable_chart import ChartCurveDTO, VariableSpeedChartDTO
 from libecalc.presentation.yaml.mappers.fluid_mapper import MEDIUM_MW_19P4
 
 StreamData = namedtuple(
@@ -26,7 +27,7 @@ CompressorData = namedtuple(
 
 
 class ProcessSimulationSingleCompressor(NamedTuple):
-    compressor_chart: dto.VariableSpeedChartDTO
+    compressor_chart: VariableSpeedChartDTO
     fluid: dto.FluidModel
     compressor: CompressorData
     inlet_stream: StreamData
@@ -40,19 +41,19 @@ def process_simulator_variable_compressor_data() -> ProcessSimulationSingleCompr
     )
 
     chart_curves = [
-        dto.ChartCurveDTO(
+        ChartCurveDTO(
             polytropic_head_joule_per_kg=[82531.50, 78440.70, 72240.80, 60105.80],
             rate_actual_m3_hour=[2900, 3504, 4003, 4595],
             efficiency_fraction=[0.72, 0.74, 0.74, 0.70],
             speed_rpm=7689,
         ),
-        dto.ChartCurveDTO(
+        ChartCurveDTO(
             polytropic_head_joule_per_kg=[135819, 129325, 121889, 110617, 98629.7],
             rate_actual_m3_hour=[3709, 4502, 4994, 5508, 5924],
             efficiency_fraction=[0.72, 0.74, 0.74, 0.73, 0.70],
             speed_rpm=9886,
         ),
-        dto.ChartCurveDTO(
+        ChartCurveDTO(
             polytropic_head_joule_per_kg=[185232, 178885, 171988, 161766, 147512, 133602],
             rate_actual_m3_hour=[4328, 4999, 5506, 6028, 6507, 6908],
             efficiency_fraction=[0.72, 0.74, 0.74, 0.74, 0.72, 0.70],
@@ -610,7 +611,7 @@ def process_simulator_variable_compressor_data() -> ProcessSimulationSingleCompr
     )
 
     test_data_set = ProcessSimulationSingleCompressor(
-        compressor_chart=dto.VariableSpeedChartDTO(curves=chart_curves),
+        compressor_chart=VariableSpeedChartDTO(curves=chart_curves),
         compressor=compressor,
         fluid=medium_fluid,
         inlet_stream=inlet_stream,
