+++ src/tests/libecalc/core/conftest.py
@@ -2,6 +2,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
+import libecalc.common.serializable_chart
 from libecalc import dto
 from libecalc.core.models.chart import SingleSpeedChart, VariableSpeedChart
 from libecalc.core.models.compressor.sampled import CompressorModelSampled
@@ -46,7 +47,7 @@ def fuel_dto() -> dto.types.FuelType:
 @pytest.fixture
 def pump_single_speed() -> PumpSingleSpeed:
     chart_curve = SingleSpeedChart(
-        dto.SingleSpeedChartDTO(
+        libecalc.common.serializable_chart.SingleSpeedChartDTO(
             rate_actual_m3_hour=[277, 524, 666, 832, 834, 927],
             polytropic_head_joule_per_kg=[10415.277000000002, 9845.316, 9254.754, 8308.089, 8312.994, 7605.693],
             efficiency_fraction=[0.4759, 0.6426, 0.6871, 0.7052, 0.7061, 0.6908],
@@ -85,7 +86,7 @@ def pump_variable_speed() -> PumpVariableSpeed:
 
     chart_curves = []
     for speed, data in df.groupby("speed"):
-        chart_curve = dto.ChartCurveDTO(
+        chart_curve = libecalc.common.serializable_chart.ChartCurveDTO(
             rate_actual_m3_hour=data["rate"].tolist(),
             polytropic_head_joule_per_kg=[x * 9.81 for x in data["head"].tolist()],  # meter liquid column to joule /kg
             efficiency_fraction=data["efficiency"].tolist(),
@@ -93,7 +94,9 @@ def pump_variable_speed() -> PumpVariableSpeed:
         )
         chart_curves.append(chart_curve)
 
-    return PumpVariableSpeed(pump_chart=VariableSpeedChart(dto.VariableSpeedChartDTO(curves=chart_curves)))
+    return PumpVariableSpeed(
+        pump_chart=VariableSpeedChart(libecalc.common.serializable_chart.VariableSpeedChartDTO(curves=chart_curves))
+    )
 
 
 @pytest.fixture
@@ -162,7 +165,7 @@ def compressor_model_sampled_3d():
 
 
 @pytest.fixture
-def variable_speed_compressor_chart_dto() -> dto.VariableSpeedChartDTO:
+def variable_speed_compressor_chart_dto() -> libecalc.common.serializable_chart.VariableSpeedChartDTO:
     df = pd.DataFrame(
         [
             [10767, 4053.0, 161345.0, 0.72],
@@ -205,7 +208,7 @@ def variable_speed_compressor_chart_dto() -> dto.VariableSpeedChartDTO:
         columns=["speed", "rate", "head", "efficiency"],
     )
     chart_curves = [
-        dto.ChartCurveDTO(
+        libecalc.common.serializable_chart.ChartCurveDTO(
             polytropic_head_joule_per_kg=data["head"].tolist(),
             rate_actual_m3_hour=data["rate"].tolist(),
             efficiency_fraction=data["efficiency"].tolist(),
@@ -214,7 +217,7 @@ def variable_speed_compressor_chart_dto() -> dto.VariableSpeedChartDTO:
         for speed, data in df.groupby("speed")
     ]
 
-    return dto.VariableSpeedChartDTO(curves=chart_curves)
+    return libecalc.common.serializable_chart.VariableSpeedChartDTO(curves=chart_curves)
 
 
 @pytest.fixture
