+++ src/tests/libecalc/core/models/chart/test_chart_curve.py
@@ -1,14 +1,14 @@
 import numpy as np
 import pytest
 
-from libecalc import dto
+from libecalc.common.serializable_chart import ChartCurveDTO
 from libecalc.core.models.chart import ChartCurve
 
 
 @pytest.fixture
 def chart_curve() -> ChartCurve:
     return ChartCurve(
-        dto.ChartCurveDTO(
+        ChartCurveDTO(
             rate_actual_m3_hour=[1, 2, 3, 4, 5],
             polytropic_head_joule_per_kg=[5, 4, 3, 2, 1],
             efficiency_fraction=[0.5, 0.6, 0.7, 0.6, 0.5],
@@ -20,7 +20,7 @@ def chart_curve() -> ChartCurve:
 @pytest.fixture
 def chart_curve_unordered() -> ChartCurve:
     return ChartCurve(
-        dto.ChartCurveDTO(
+        ChartCurveDTO(
             rate_actual_m3_hour=[2, 3, 4, 5, 1],
             polytropic_head_joule_per_kg=[4, 3, 2, 1, 5],
             efficiency_fraction=[0.6, 0.7, 0.6, 0.5, 0.5],
@@ -54,7 +54,7 @@ def test_chart_curve_data_add_valid_and_invalid_curve_data_point(chart_curve, ch
 
 def test_chart_curve_100_percent_efficient():
     chart_curve_100_percent_efficient = ChartCurve(
-        dto.ChartCurveDTO(
+        ChartCurveDTO(
             rate_actual_m3_hour=[1, 2], polytropic_head_joule_per_kg=[2, 1], efficiency_fraction=[1, 1], speed_rpm=1
         )
     )
@@ -66,7 +66,7 @@ def test_chart_curve_data_invalid_setup_from_arrays(caplog):
     caplog.set_level("CRITICAL")
     with pytest.raises(ValueError):
         ChartCurve(
-            dto.ChartCurveDTO(
+            ChartCurveDTO(
                 rate_actual_m3_hour=[1, 2.0],
                 polytropic_head_joule_per_kg=[3, 4.0],
                 efficiency_fraction=[30.0, 70.0],
@@ -77,7 +77,7 @@ def test_chart_curve_data_invalid_setup_from_arrays(caplog):
     # Chart curve should have at least two points
     with pytest.raises(ValueError) as ve:
         ChartCurve(
-            dto.ChartCurveDTO(
+            ChartCurveDTO(
                 rate_actual_m3_hour=[1],
                 polytropic_head_joule_per_kg=[3],
                 efficiency_fraction=[0.3],
@@ -90,7 +90,7 @@ def test_chart_curve_data_invalid_setup_from_arrays(caplog):
 def test_chart_curve_sort_input_values():
     """Values should be sorted by the root validator in pydantic."""
     curve = ChartCurve(
-        dto.ChartCurveDTO(
+        ChartCurveDTO(
             rate_actual_m3_hour=[3, 2, 4, 1],
             polytropic_head_joule_per_kg=[1, 2, 3, 4],
             efficiency_fraction=[1, 1, 1, 1],
@@ -110,7 +110,7 @@ def test_efficiency_as_function_of_rate(chart_curve):
 
 def test_head_as_function_of_rate(chart_curve):
     _ = ChartCurve(
-        dto.ChartCurveDTO(
+        ChartCurveDTO(
             rate_actual_m3_hour=[1, 2, 3, 4, 5],
             polytropic_head_joule_per_kg=[5, 4, 3, 2, 1],
             efficiency_fraction=[0.5, 0.6, 0.7, 0.6, 0.5],
@@ -146,7 +146,7 @@ def test_rate_head_and_efficiency_at_maximum_rate(chart_curve):
 def test_distance_efficiency():
     """Integration test."""
     chart_curve = ChartCurve(
-        dto.ChartCurveDTO(
+        ChartCurveDTO(
             rate_actual_m3_hour=[577, 336, 708, 842, 824, 826, 825, 1028],
             polytropic_head_joule_per_kg=[1718.7, 1778.7, 1665.2, 1587.8, 1601.9, 1601.9, 1602.7, 1460.6],
             efficiency_fraction=[0.6203, 0.4717, 0.6683, 0.6996, 0.695, 0.6975, 0.6981, 0.7193],
@@ -170,7 +170,7 @@ def test_interpolation_functions():
     # NB: Minimum rate not first, will be sorted
     # Random order of input columns
     chart_curve = ChartCurve(
-        dto.ChartCurveDTO(
+        ChartCurveDTO(
             rate_actual_m3_hour=[577, 336, 708, 842, 824, 826, 825, 1028],
             polytropic_head_joule_per_kg=[1718.7, 1778.7, 1665.2, 1587.8, 1601.9, 1601.9, 1602.7, 1460.6],
             efficiency_fraction=[0.6203, 0.4717, 0.6683, 0.6996, 0.695, 0.6975, 0.6981, 0.7193],
