+++ src/tests/libecalc/input/mappers/test_consumer_chart.py
@@ -1,6 +1,7 @@
 import pytest
 
 from libecalc import dto
+from libecalc.common.serializable_chart import SingleSpeedChartDTO
 from libecalc.common.units import Unit
 
 # from libecalc.presentation.yaml.mappers import _single_speed_compressor_chart_mapper
@@ -85,7 +86,7 @@ class TestSingleSpeedChart:
             facility_data=pump_chart,
         )
         assert pump_model_dto == dto.PumpModel(
-            chart=dto.SingleSpeedChartDTO(
+            chart=SingleSpeedChartDTO(
                 rate_actual_m3_hour=[6.0, 6.0],
                 polytropic_head_joule_per_kg=[
                     Unit.POLYTROPIC_HEAD_METER_LIQUID_COLUMN.to(Unit.POLYTROPIC_HEAD_JOULE_PER_KG)(x)
@@ -105,7 +106,7 @@ class TestSingleSpeedChart:
             facility_data=pump_chart,
         )
         assert pump_model_dto == dto.PumpModel(
-            chart=dto.SingleSpeedChartDTO(
+            chart=SingleSpeedChartDTO(
                 rate_actual_m3_hour=[6.0, 6.0],
                 polytropic_head_joule_per_kg=[
                     Unit.POLYTROPIC_HEAD_METER_LIQUID_COLUMN.to(Unit.POLYTROPIC_HEAD_JOULE_PER_KG)(x)
@@ -150,7 +151,7 @@ class TestCompressorChartSingleSpeed:
         chart_dto = _single_speed_compressor_chart_mapper(
             model_config=compressor_chart, resources={"compressorchart.csv": chart_resource_with_speed}
         )
-        assert chart_dto == dto.SingleSpeedChartDTO(
+        assert chart_dto == SingleSpeedChartDTO(
             rate_actual_m3_hour=[6.0, 6.0],
             polytropic_head_joule_per_kg=[7000.0, 7000.0],
             efficiency_fraction=[0.08, 0.08],
@@ -162,7 +163,7 @@ class TestCompressorChartSingleSpeed:
         chart_dto = _single_speed_compressor_chart_mapper(
             model_config=compressor_chart, resources={"compressorchart.csv": chart_resource_without_speed}
         )
-        assert chart_dto == dto.SingleSpeedChartDTO(
+        assert chart_dto == SingleSpeedChartDTO(
             rate_actual_m3_hour=[6.0, 6.0],
             polytropic_head_joule_per_kg=[7000.0, 7000.0],
             efficiency_fraction=[0.08, 0.08],
